Move header file and fix comment
